Restructure Event views
