+++ src/views/Event/Details.vue
